pyproject.toml
[build-system]
requires = ["setuptools>=63", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prisma-tools"
version = "0.0.2"
description = "Tools for Prisma Access (Cloud Managed)"
readme = "README.md"
dependencies = ["typer==0.7.0","requests==2.28.2","lxml==4.9.2", "rich==13.3.5", "ruamel.yaml==0.17.26"]
requires-python = ">=3.10"
authors = [{ name = "Ryan Gillespie", email = "rgillespie@gmail.com" }]
keywords = ["automation", "infrastructure", "network"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking :: Firewalls",
    "Topic :: Utilities",
]
[project.optional-dependencies]
dev = ["pytest==7.2.1","black==22.12.0", "isort==5.12.0", "pylama==8.4.1", "mypy==1.2.0"]

[project.scripts]
ptools = "prisma_tools.cli:app"

[tool.isort]
profile = "black"
skip = [".venv/", "prisma_tools/PrismaSASECloudManaged_Python/", "tests/"]

[tool.black]
include = "prisma_tools/\\S+\\.py$"
exclude = "prisma_tools/PrismaSASECloudManaged_Python"

[tool.mypy]
python_version = "3.10"
pretty = true
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_configs = true
strict_optional = true
exclude = [".venv/", "tests/", "prisma_tools/PrismaSASECloudManaged_Python/"]

[tool.pylama]
max-line-length = 120
skip = ".venv/*,tests/*,prisma_tools/PrismaSASECloudManaged_Python/*"
ignore = ['D101', 'D202', 'D203', 'D212', 'D400', 'D406', 'D407', 'D408', 'D409', 'D415']
convention = "google"

[tool.pydocstyle]
match_dir = "^prisma_tools/*"
ignore = ['D101', 'D202', 'D203', 'D212', 'D400', 'D406', 'D407', 'D408', 'D409', 'D415']
# D101: missing docstring in public class
# D202: No blank lines allowed after function docstring
# D203: 1 blank line required before class docstring
# D212: Multi-line docstring summary should start at the first line
# D400: First line should end with a period
# D406: Section name should end with a newline
# D407: Missing dashed underline after section
# D408: Section underline should be in the line following the sections name
# D409: Section underline should match the length of its name
# D415: first line should end with a period, question mark, or exclamation point
